// src/lib/fineract/generated/types.gen.ts
export type ExternalId = {
    empty?: boolean;
    value?: string;
};
// ...
export type GetRolesResponse = {
    description?: string;
    id?: number;
    name?: string;
};
// ...
export type GetUsersResponse = {
    email?: string;
    firstname?: string;
    id?: number;
    lastname?: string;
    officeId?: number;
    officeName?: string;
    passwordNeverExpires?: boolean;
    selectedRoles?: Array<RoleData>;
    staff?: StaffData;
    username?: string;
};
// ...
export type Image = {
    id?: number;
    location?: string;
    new?: boolean;
    storageType?: number;
};
// ...
export type Office = {
    children?: Array<Office>;
    externalId?: ExternalId;
    hierarchy?: string;
    id?: number;
    name?: string;
    new?: boolean;
    openingDate?: string;
    openingLocalDate?: string;
    parent?: Office;
};

export type OfficeData = {
    dateFormat?: string;
    externalId?: ExternalId;
    hierarchy?: string;
    id?: number;
    locale?: string;
    name?: string;
    nameDecorated?: string;
    openingDate?: string;
    parentId?: number;
    parentName?: string;
    rowIndex?: number;
};
// ...
export type RoleData = {
    id?: number;
    name?: string;
};
// ...
export type Staff = {
    active?: boolean;
    displayName?: string;
    emailAddress?: string;
    externalId?: string;
    firstname?: string;
    id?: number;
    image?: Image;
    joiningDate?: string;
    lastname?: string;
    loanOfficer?: boolean;
    mobileNo?: string;
    new?: boolean;
    notActive?: boolean;
    notLoanOfficer?: boolean;
    office?: Office;
    organisationalRoleParentStaff?: Staff;
    organisationalRoleType?: number;
};

export type StaffAssignmentHistory = {
    createdBy?: number;
    createdDate?: string;
    currentRecord?: boolean;
    id?: number;
    lastModifiedBy?: number;
    lastModifiedDate?: string;
    new?: boolean;
    startDate?: string;
};

export type StaffData = {
    allowedOffices?: Array<OfficeData>;
    dateFormat?: string;
    displayName?: string;
    externalId?: string;
    firstname?: string;
    id?: number;
    isActive?: boolean;
    isLoanOfficer?: boolean;
    joiningDate?: string;
    lastname?: string;
    locale?: string;
    mobileNo?: string;
    officeId?: number;
    officeName?: string;
    rowIndex?: number;
};